backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==21.12b0
click==8.0.3
Django==4.0.1
django-formtools==2.3
django-otp==1.1.3
django-phonenumber-field==6.1.0
django-phonenumber-field[phonenumberslite]==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg==3.0.7
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
qrcode==7.3.1
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2021.5
PyCryptodome==3.14.1
